requirements.txt
scipy
tqdm
torch
torchvision
monai
nibabel
tensorboard
scikit-image
scikit-learn
pyyaml
einops
SimpleITK